Update keyword argument processing to ruby 3 style
The core CSV.parse and CSV.generate use keyword arguments, so passing a
hash no longer works on ruby 3.0.  The double splat operator came out
back in Ruby 2.0, so this new code should work on all ruby versions we
support.

The semantics are subtly different in that the old version accepted any
form of arguments whereas the new version will blow up if the wrong
number of arguments is passed, or if the second parameter doesn't
respond to merge! However, I don't think this would break any code that
was actually working.